requirements.txt
aiofiles==22.1.0
anytree==2.8.0
beautifulsoup4==4.11.1
certifi==2022.9.24
chardet==4.0.0
click==8.0.1
dnspython==2.2.1
fastapi==0.67.0
h11==0.14.0
idna==3.2
Jinja2==3.0.1
karton-core==4.2.0
liccheck==0.7.3
MarkupSafe==2.1.1
minio==7.1.12
psycopg2-binary==2.9.1
pydantic==1.10.2
pymongo==3.12.0
python-multipart==0.0.5
redis==3.5.3
requests==2.26.0
requests-mock==1.10.0
shodan==1.25.0
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.6
uvicorn==0.20.0
